Remove lambda auto argument type detection.   Remove constexpr from fail_fast in gsl due to Linux not being happy.   Comment out std::stream support due to MacOS std lib broken.

+++ onnxruntime/core/providers/cpu/math/element_wise_ops.h
@@ -321,8 +321,8 @@ struct BroadcastIterator {
   }
 
   void Reserve(int64_t max_dims) {
-    deltas_.reserve(max_dims);
-    counts_.reserve(max_dims);
+    deltas_.reserve(static_cast<size_t>(max_dims));
+    counts_.reserve(static_cast<size_t>(max_dims));
   }
 
   void Init(int64_t axis, int64_t largest) {
